backend/app/user/schemas.py
from pydantic import BaseModel, field_validator
from pydantic_core.core_schema import FieldValidationInfo


class UserSchema(BaseModel):
    email: str
    password: str


class UserRegister(UserSchema):
    confirm_password: str

    @field_validator("password")
    def password_valid(cls, v, info: FieldValidationInfo):
        if "password" in info.data and 8 > len(v) > 100:
            raise ValueError("password not valid")
        return v

    @field_validator("confirm_password")
    def passwords_match(cls, v, info: FieldValidationInfo):
        if "password" in info.data and v != info.data["password"]:
            raise ValueError("passwords do not match")
        return v
