backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["yehor <egor.karabanov@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.35"}
alembic = "^1.13.2"
loguru = "^0.7.2"
celery = {extras = ["redis"], version = "^5.4.0"}
flower = "^2.0.1"
pydantic-settings = "^2.5.2"
ruff = "^0.6.9"
bcrypt = "^4.2.0"
itsdangerous = "^2.2.0"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
